Decouple IndexSettings from IncludeExclude (#2860)
This change refactors an earlier change to impose a reg-ex size limit on the include/exclude string. Instead of accepting an IndexSettings instance, the class now accepts a integer limit value. This is necessary because the IncludeExclude class is used outside the core codebase, whose use-cases may be unaware of indices and their settings. To ensure that a limit is always imposed, a default limit is defined in the class.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>
(cherry picked from commit ba1966853e728b153e42be59ba449420e79b09ee)